tmux-asciicast-pane: Fix asciicast path.
Since `tmux pipe-pane` runs commands on the server, they don't have
access to the working directory. Thus we need to work out the full path
of the script.

This uses realpath/bash since it's quite clean approach for Linux. A
portable version could be written, but would be substantially more
complex.